+++ changes.md
@@ -6,7 +6,7 @@
 3. Add support of TGeoTessellated shape
 4. Major changes in v7 drawing: RFrame, RPalette, RColor, RStatBox, ...
 5. Fix in reading std::map member-wise
-6. Better handling of context menu position 
+6. Better handling of context menu position
 
 
 ## Changes in 5.8.1